+++ java-calculator/src/main/java/com/programmers/junho/domain/Operator.java
@@ -2,5 +2,5 @@ package com.programmers.junho.domain;
 
 @FunctionalInterface
 public interface Operator {
-    double apply(double a, double b);
+    int apply(int a, int b);
 }
