+++ java-calculator/src/main/java/com/programmers/junho/view/OutputView.java
@@ -7,5 +7,5 @@ public interface OutputView {
     void printChoiceMessage();
     void printExpressions(List<String> expressions);
 
-    void printCalculatedResult(double result);
+    void printCalculatedResult(int result);
 }
